+++ R/full_factor.R
@@ -166,7 +166,7 @@ summary.full_factor <- function(
 
   cat("Factor analysis\n")
   cat("Data        :", object$df_name, "\n")
-  if (!is_empty(object$data_filter)) {
+  if (!radiant.data::is_empty(object$data_filter)) {
     cat("Filter      :", gsub("\\n", "", object$data_filter), "\n")
   }
   cat("Variables   :", paste0(object$vars, collapse = ", "), "\n")
@@ -246,7 +246,7 @@ summary.full_factor <- function(
 #' @param x Return value from \code{\link{full_factor}}
 #' @param plots Include attribute ("attr"), respondents ("resp") or both in the plot
 #' @param shiny Did the function call originate inside a shiny app
-#' @param custom Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and \url{http://docs.ggplot2.org} for options.
+#' @param custom Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and \url{https://ggplot2.tidyverse.org/} for options.
 #' @param ... further arguments passed to or from other methods
 #'
 #' @examples
@@ -334,7 +334,7 @@ plot.full_factor <- function(x, plots = "attr", shiny = FALSE, custom = FALSE, .
 #'
 #' @export
 store.full_factor <- function(dataset, object, name = "", ...) {
-  if (is_empty(name)) name <- "factor"
+  if (radiant.data::is_empty(name)) name <- "factor"
   fscores <- as.data.frame(object$fres$scores, stringsAsFactors = FALSE)
   indr <- indexr(dataset, object$vars, object$data_filter)
   fs <- data.frame(matrix(NA, nrow = indr$nr, ncol = ncol(fscores)), stringsAsFactors = FALSE)