+++ R/pre_factor.R
@@ -12,7 +12,7 @@
 #'
 #' @examples
 #' pre_factor(shopping, "v1:v6") %>% str()
-#'
+#' 
 #' @seealso \code{\link{summary.pre_factor}} to summarize results
 #' @seealso \code{\link{plot.pre_factor}} to plot results
 #'
@@ -51,21 +51,22 @@ pre_factor <- function(dataset, vars, hcor = FALSE, data_filter = "", envir = pa
   pre_kmo <- psych::KMO(cmat)
   pre_eigen <- eigen(cmat)$values
 
-  err_mess <- "The selected variables are perfectly collinear. Please check the correlations\nand remove any variable with a correlation of 1 or -1 from the analysis"
   if (det(cmat) > 0) {
     scmat <- try(solve(cmat), silent = TRUE)
     if (inherits(scmat, "try-error")) {
-      pre_r2 <- err_mess
+      pre_r2 <- matrix(NA, nrow = nrow(cmat), ncol = 1)
+      rownames(pre_r2) <- rownames(cmat)
     } else {
       pre_r2 <- {1 - (1 / diag(scmat))} %>%
         data.frame(stringsAsFactors = FALSE) %>%
         set_colnames("Rsq")
     }
   } else {
-    pre_r2 <- err_mess
+    pre_r2 <- matrix(NA, nrow = nrow(cmat), ncol = 1)
+    rownames(pre_r2) <- rownames(cmat)
   }
 
-  rm(dataset, err_mess, envir)
+  rm(dataset, envir)
 
   as.list(environment()) %>% add_class("pre_factor")
 }
@@ -91,7 +92,6 @@ pre_factor <- function(dataset, vars, hcor = FALSE, data_filter = "", envir = pa
 #' result <- pre_factor(shopping, "v1:v6")
 #' summary(result)
 #' pre_factor(computer, "high_end:business") %>% summary()
-#'
 #' @seealso \code{\link{pre_factor}} to calculate results
 #' @seealso \code{\link{plot.pre_factor}} to plot results
 #'
@@ -106,7 +106,7 @@ summary.pre_factor <- function(object, dec = 2, ...) {
 
   cat("Pre-factor analysis diagnostics\n")
   cat("Data        :", object$df_name, "\n")
-  if (!is_empty(object$data_filter)) {
+  if (!radiant.data::is_empty(object$data_filter)) {
     cat("Filter      :", gsub("\\n", "", object$data_filter), "\n")
   }
   cat("Variables   :", paste0(object$vars, collapse = ", "), "\n")
@@ -138,9 +138,8 @@ summary.pre_factor <- function(object, dec = 2, ...) {
   cat("Bartlett test\n")
   cat("Null hyp. : variables are not correlated\n")
   cat("Alt. hyp. : variables are correlated\n")
-  bt <- object$btest$p.value %>% {
-    if (. < .001) "< .001" else round(., dec + 1)
-  }
+  bt <- object$btest$p.value
+  bt <- if (!radiant.data::is_empty(bt) && bt < .001) "< .001" else round(bt, dec + 1)
   cat(paste0(
     "Chi-square: ", round(object$btest$chisq, 2), " df(",
     object$btest$df, "), p.value ", bt, "\n"
@@ -151,6 +150,9 @@ summary.pre_factor <- function(object, dec = 2, ...) {
   # print(object$pre_kmo$MSAi, digits = dec)
 
   cat("\nVariable collinearity:\n")
+  if (all(is.na(object$pre_r2))) {
+    cat("\nRsq measures could not be calculated because the selected variables\nare perfectly collinear. Please check the correlations and remove\nany variable with a correlation of 1 or -1 from the analysis\n\n")
+  }
   data.frame(Rsq = object$pre_r2, KMO = object$pre_kmo$MSAi, stringsAsFactors = FALSE) %>%
     format_df(dec = dec) %>%
     set_rownames(rownames(object$pre_r2)) %>%
@@ -180,13 +182,12 @@ summary.pre_factor <- function(object, dec = 2, ...) {
 #' @param plots Plots to return. "change" shows the change in eigenvalues as variables are grouped into different number of factors, "scree" shows a scree plot of eigenvalues
 #' @param cutoff For large datasets plots can take time to render and become hard to interpret. By selection a cutoff point (e.g., eigenvalues of .8 or higher) factors with the least explanatory power are removed from the plot
 #' @param shiny Did the function call originate inside a shiny app
-#' @param custom Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and \url{http://docs.ggplot2.org} for options.
+#' @param custom Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and \url{https://ggplot2.tidyverse.org/} for options.
 #' @param ... further arguments passed to or from other methods
 #'
 #' @examples
 #' result <- pre_factor(shopping, "v1:v6")
 #' plot(result, plots = c("change", "scree"), cutoff = .05)
-#'
 #' @seealso \code{\link{pre_factor}} to calculate results
 #' @seealso \code{\link{summary.pre_factor}} to summarize results
 #'
@@ -206,11 +207,11 @@ plot.pre_factor <- function(
   if ("scree" %in% plots) {
     plot_list[[which("scree" == plots)]] <-
       ggplot(dat, aes(x = x, y = y, group = 1)) +
-        geom_line(color = "blue", linetype = "dotdash", size = .7) +
-        geom_point(color = "blue", size = 4, shape = 21, fill = "white") +
-        geom_hline(yintercept = 1, color = "black", linetype = "solid", size = .5) +
-        labs(title = "Screeplot", x = "# factors", y = "Eigenvalues") +
-        scale_x_continuous(breaks = dat[["x"]])
+      geom_line(color = "blue", linetype = "dotdash", size = .7) +
+      geom_point(color = "blue", size = 4, shape = 21, fill = "white") +
+      geom_hline(yintercept = 1, color = "black", linetype = "solid", size = .5) +
+      labs(title = "Screeplot", x = "# factors", y = "Eigenvalues") +
+      scale_x_continuous(breaks = dat[["x"]])
   }
 
   if ("change" %in% plots) {