+++ inst/app/tools/analysis/kclus_ui.R
@@ -103,10 +103,11 @@ output$ui_kclus <- renderUI({
         ),
         conditionalPanel(
           condition = "input.km_vars != null",
-          HTML("<label>Store cluster membership:</label>"),
+          # HTML("<label>Store cluster membership:</label>"),
+          tags$label("Store cluster membership:"),
           tags$table(
             tags$td(uiOutput("ui_km_store_name")),
-            tags$td(actionButton("km_store", "Store", icon = icon("plus")), style = "padding-top:5px;")
+            tags$td(actionButton("km_store", "Store", icon = icon("plus")), class = "top_mini")
           )
         )
       ),
@@ -128,7 +129,7 @@ output$ui_kclus <- renderUI({
 })
 
 km_plot <- eventReactive(c(input$km_run, input$km_plots), {
-  if (.km_available() == "available" && !is_empty(input$km_plots, "none")) {
+  if (.km_available() == "available" && !radiant.data::is_empty(input$km_plots, "none")) {
     list(plot_width = 750, plot_height = 300 * ceiling(length(input$km_vars) / 2))
   }
 })
@@ -197,7 +198,7 @@ output$kclus <- renderUI({
 .plot_kclus <- eventReactive(c(input$km_run, input$km_plots), {
   if (.km_available() != "available") {
     .km_available()
-  } else if (is_empty(input$km_plots, "none")) {
+  } else if (radiant.data::is_empty(input$km_plots, "none")) {
     "Please select a plot type from the drop-down menu"
   } else {
     withProgress(message = "Generating plots", value = 1, {
@@ -208,7 +209,7 @@ output$kclus <- renderUI({
 
 observeEvent(input$kclus_report, {
   inp_out <- list(list(dec = 2), "")
-  if (!is_empty(input$km_plots, "none")) {
+  if (!radiant.data::is_empty(input$km_plots, "none")) {
     figs <- TRUE
     outputs <- c("summary", "plot")
     inp_out[[2]] <- list(plots = input$km_plots, custom = FALSE)
@@ -217,7 +218,7 @@ observeEvent(input$kclus_report, {
     figs <- FALSE
   }
 
-  if (!is_empty(input$km_store_name)) {
+  if (!radiant.data::is_empty(input$km_store_name)) {
     fixed <- fix_names(input$km_store_name)
     updateTextInput(session, "km_store_name", value = fixed)
     xcmd <- glue('{input$dataset} <- store({input$dataset}, result, name = "{fixed}")')